Moved return url calculation to FileAdmin/FolderAdmin the common ancestor
