+++ filer/admin/permissions.py
@@ -1,5 +1,6 @@
 #-*- coding: utf-8 -*-
 from django.contrib import admin
+from django.core.urlresolvers import reverse
 
 
 class PrimitivePermissionAwareModelAdmin(admin.ModelAdmin):
@@ -21,3 +22,13 @@ class PrimitivePermissionAwareModelAdmin(admin.ModelAdmin):
     def has_delete_permission(self, request, obj=None):
         # we don't have a specific delete permission... so we use change
         return self.has_change_permission(request, obj)
+
+    def _get_post_url(self, obj):
+        """ Needed to retrieve the changelist url as Folder/File cna be extended
+        and admin url may change """
+        ## Code borrowed from django ModelAdmin to determine changelist on the fly
+        opts = obj._meta
+        module_name = opts.module_name
+        return reverse('admin:%s_%s_changelist' %
+                       (opts.app_label, module_name),
+            current_app=self.admin_site.name)
